put function in seperate file
